Consolidate stereo solver tests into single file; remove internal scripts
- Merge 6 phase test files into one tests/solvers/test_cp_local_search_stereo_insertion.py
- Remove internal 'phase' references from test docstrings
- Remove solvers/.../scripts/validate_repeatability.py intermediate script
- Remove empty scripts/ directory
- All 50 tests pass
